# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Evaluation Service API
  version: 1.2.0
  description: |
    Full API contract for the current backend implementation (Phases 1-5).
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/assignments:
    get:
      tags: [Campaigns]
      summary: List rich assignments (admin)
      parameters:
        - in: query
          name: campaignId
          schema: { type: string }
        - in: query
          name: stepType
          schema: { type: string }
        - in: query
          name: sectionId
          schema: { type: string }
        - in: query
          name: facultyId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: evaluatorId
          schema: { type: string }
        - in: query
          name: evaluateeId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
        - in: query
          name: sortBy
          schema: { type: string, default: updatedAt }
        - in: query
          name: sortDir
          schema: { type: string, default: desc }
      responses:
        '200':
          description: Assignment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListResponse'
    post:
      tags: [Campaigns]
      summary: Create rich assignment (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '409':
          description: Duplicate assignment tuple conflict
          content:
            application/json:
              schema:
                type: object
              examples:
                duplicateAssignment:
                  value:
                    type: https://api.evaluationservice.com/errors/duplicate-assignment
                    title: Duplicate Assignment
                    status: 409
                    detail: Duplicate assignment tuple for campaign/evaluator/evaluatee/role
                    campaignId: c-001
                    evaluatorId: e-100
                    evaluateeId: u-200
                    evaluatorRole: PEER
                    existingAssignmentId: a-existing
  /api/v1/assignments/{id}:
    get:
      tags: [Campaigns]
      summary: Get assignment by id (admin)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
    put:
      tags: [Campaigns]
      summary: Update assignment metadata (admin)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssignmentRequest'
      responses:
        '200':
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
...
    CreateAssignmentRequest:
      type: object
      required: [campaignId, evaluatorId, evaluateeId, evaluatorRole]
      properties:
        campaignId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        evaluatorRole: { type: string }
        stepType: { type: string, nullable: true }
        sectionId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        anonymityMode: { type: string, nullable: true }
        status: { type: string, nullable: true }
    UpdateAssignmentRequest:
      type: object
      properties:
        stepType: { type: string, nullable: true }
        sectionId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        anonymityMode: { type: string, nullable: true }
        status: { type: string, nullable: true }
...
    AssignmentResponse:
      type: object
      properties:
        id: { type: string }
        campaignId: { type: string }
        evaluatorId: { type: string }
        evaluateeId: { type: string }
        evaluatorRole: { type: string }
        completed: { type: boolean }
        evaluationId: { type: string, nullable: true }
        stepType: { type: string, nullable: true }
        sectionId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        anonymityMode: { type: string, nullable: true }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssignmentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentResponse'
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
...